% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_effectiveness_table.R
\name{cost_effectiveness_table.bcea}
\alias{cost_effectiveness_table.bcea}
\title{cost_effectiveness_table.bcea}
\usage{
\method{cost_effectiveness_table}{bcea}(dat, wtp_threshold = 20000)
}
\arguments{
\item{wtp_threshold}{}
}
\description{
cost_effectiveness_table.bcea
}
\examples{

}
