% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_effectiveness_table.R
\name{cost_effectiveness_table.default}
\alias{cost_effectiveness_table.default}
\title{cost_effectiveness_table.default}
\usage{
\method{cost_effectiveness_table}{default}(dat, html = TRUE)
}
\arguments{
\item{dat}{cost and health realisations}

\item{html}{output format}
}
\description{
cost_effectiveness_table.default
}
\examples{

}
