% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree.R
\name{dectree}
\alias{dectree}
\title{Decision Tree Calculation}
\usage{
dectree(
  data,
  nsim = 1000,
  costDistns,
  time_res,
  drug = drug,
  prevalence = 0.25,
  performance,
  c.newtest = 0,
  name.newtest = NA,
  quant = 0.5,
  QALYloss,
  N = nrow(data),
  wholecohortstats = FALSE,
  terminal_health,
  terminal_cost,
  followup_pdf,
  ...
)
}
\arguments{
\item{data}{IDEA study data (data.frame)}

\item{nsim}{Number of sample points (integer) Default: 1000}

\item{costDistns}{List of distribution names and parameter values for each test/procedure}

\item{time_res}{time to obtain novel test result (list)}

\item{drug}{treatment drug cost and quantities (list)}

\item{prevalence}{TB proportion in cohort (double 0-1) Default: 0.25}

\item{performance}{Sensitivity and specificity test (list)}

\item{c.newtest}{Rule-out test unit cost (double) Defalut: 0}

\item{name.newtest}{Name of rule-out test to get at distribution (string)}

\item{quant}{Quantile value of time to diagnosis and costs from observed data (double 0-1) Default: 0.5 (median)}

\item{QALYloss}{QALY loss due to active TB (list)}

\item{N}{Number of patients. The number in the data is used as default}

\item{wholecohortstats}{Should the output stats be the total or per patient}

\item{terminal_health}{function of terminal node total QALY loss}

\item{terminal_cost}{function of terminal node total costs}

\item{followup_pdf}{Follow-up appointment time probability distribution function}

\item{...}{}
}
\value{
Health and cost realisations (list)
}
\description{
Probabilistically incorporates sampling variability
(and standard pathway cost and time to diagnosis).
}
\examples{

}
