% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dectree2.R
\name{dectree2}
\alias{dectree2}
\title{Decision Tree Calculation}
\usage{
dectree2(data, nsim = 100, costDistns, time_res, prevalence = 0.25,
  cutoff = 1, FNtime = 42, FNdist = TRUE, performance, c.ruleout = 0,
  name.ruleout = NA, quant = 0.5, utility = NA, N = nrow(data),
  wholecohortstats = FALSE, terminal_health, terminal_cost)
}
\arguments{
\item{data}{IDEA study data (data.frame)}

\item{nsim}{Number of sample points (integer) Default: 1000}

\item{costDistns}{List of distribution names and parameter values for each test/procedure}

\item{time_res}{time to obtain novel test result (list)}

\item{prevalence}{TB proportion in cohort (double 0-1) Default: 0.25}

\item{cutoff}{Clinical judgement threshold (double 0-1) Default: 1}

\item{FNtime}{False negative follow-up time fixed value (double) Default: 42}

\item{FNdist}{Should false negative time to follow-up distribution be used (logical) Default: TRUE}

\item{performance}{Sensitivity and specificity test (list)}

\item{c.ruleout}{Rule-out test unit cost (double) Defalut: 0}

\item{name.ruleout}{Name of rule-out test to get at distribution (string)}

\item{quant}{Quantile value of time to diagnosis and costs from observed data (double 0-1) Default: 0.5 (median)}

\item{utility}{QALY adjustment utility due to active TB}

\item{N}{Number of patients. The number in the data is used as default}

\item{wholecohortstats}{Should the output stats be the total or per patient}

\item{terminal_health}{function of terminal node total QALY loss}

\item{terminal_cost}{function of terminal node total costs}
}
\value{
Health and cost realisations (list)
}
\description{
Probabilistically incorporates sampling variability
(and standard pathway cost and time to diagnosis).
}
\examples{

}
