% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_sample.R
\name{inverse_sample}
\alias{inverse_sample}
\title{Inverse Sample}
\usage{
inverse_sample(cdf)
}
\arguments{
\item{cdf}{Cumulative density function}
}
\value{
sampled realisation
}
\description{
Inverse Sample
}
\examples{

}
