% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_drug_cost.R
\name{total_drug_cost}
\alias{total_drug_cost}
\title{Calculate Total Drug Costs Per Day}
\usage{
total_drug_cost(drug, weight = 1)
}
\arguments{
\item{weight}{}
}
\value{

}
\description{
Calculate Total Drug Costs Per Day
}
